Remove INHERITED_WEB from here (needs to be core patch)
Unfortunately this will not work. This way the settings are only
injected, when they are loaded for the first time. Afterwards Prefs will
use its cache and the preferences are missing in the cache.

Instead this will be done via a core patch. Prefs.pm will utilize the
inheritsFrom method then.

Ticket: none
Release-Note: none